Move LayoutData::push_run into the shape module
